Explicit casts of mask

+++ TeXnicle/TeXEditor/TextView/TPSectionListController.m
@@ -192,7 +192,7 @@ NSString *TPsectionListPopupTitle = @"Jump to section...";
 	
 	NSEvent *event =  [NSEvent mouseEventWithType:NSLeftMouseDown
 																			 location:menuOrigin
-																	modifierFlags:NSLeftMouseDownMask // 0x100
+																	modifierFlags:(int)NSLeftMouseDownMask // 0x100
 																			timestamp:0
 																	 windowNumber:[[(NSButton *)sender window] windowNumber]
 																				context:[[(NSButton *)sender window] graphicsContext]
